move from components/pos to src/pages/PointOfSale
