+++ src/pages/PointOfSale/ItemList.vue
@@ -2,7 +2,7 @@
     <div class="row">
         <button class="list-group-item col-sm-4" v-for="item in items" :key="item.name" @click="itemClicked(item)"> 
             <strong>{{ item.name }}</strong><br>
-            <img src="../images/mobile.jpeg"><br>
+            <img src="./images/mobile.jpeg"><br>
             <strong>INR</strong> - {{ item.rate }}
         </button>
     </div>
